% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aal_border.R
\docType{data}
\name{aal_border}
\alias{aal_border}
\title{AAL List of Border Vertices}
\format{A list with 2 elements, which are either `v4` or `v5`,
each having a `list` of elements, :
\describe{
\item{BORDER.V}{a numeric vector of values corresponding to those
in the image}
\item{BORDER.XYZ}{a V by 3 matrix of vertices}
}}
\usage{
aal_border
}
\description{
AAL List of Border Vertices
}
\keyword{datasets}
