% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aal_colors.R
\docType{data}
\name{aal_colors}
\alias{aal_colors}
\title{AAL Colors}
\format{A vector of colors that are in RGB.  There is no blank color
for background as \code{\link{ortho2}} usually removes background.}
\usage{
aal_colors
}
\description{
AAL Colors
}
\keyword{datasets}
