% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aal_fname.R
\name{aal_fname}
\alias{aal_fname}
\alias{aal_image}
\title{Filename of AAL image}
\usage{
aal_fname(version = c("5", "4"))

aal_image(version = c("5", "4"))
}
\arguments{
\item{version}{Version of AAL ROI image, versions 4 and 5.}
}
\value{
A character vector
}
\description{
Filename of AAL image
}
\examples{
aal_fname(version = 4)
aal_fname(version = 5)
aal_fname(version = "4")
aal_fname(version = "5")

v4_img = aal_image(version = 4)
v5_img = aal_image()

}
