% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aal_labels.R
\docType{data}
\name{aal_labels}
\alias{aal_labels}
\title{AAL Label Names}
\format{A list with 2 elements, which are either `v4` or `v5`,
each having a `data.frame` of:
\describe{
\item{index}{the corresponding value of the image}
\item{name}{the name of the region of the brain}
}}
\usage{
aal_labels
}
\description{
AAL Label Names
}
\keyword{datasets}
