% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aal_lookup.R
\name{aal_lookup}
\alias{aal_lookup}
\title{Lookup specific indices for labels}
\usage{
aal_lookup(xyz, version = c("5", "4"))
}
\arguments{
\item{xyz}{Indices in coordinates of the atlas to look up specific}

\item{version}{Version of AAL ROI image, versions 4 and 5.}
}
\value{
A \code{data.frame} with the results
}
\description{
Lookup specific indices for labels
}
\examples{
aal_lookup(xyz = c(61, 60, 67))
aal_lookup(xyz = rbind(c(61, 60, 67),
c(50, 44, 32)
))
}
